branch channel roles into dedicated helper
